Rewrite with simpler API. Apply it to Content as well
